Remove welcome bonus logic and refresh main menu tests
